refactor(cmd): move repl to root command
