+++ internal/cmd/deprecated.go
@@ -1,0 +1,17 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+func init() {
+	rootCmd.AddCommand(&cobra.Command{
+		Use:   "repl",
+		Short: "deprecated: use `gh-not --repl` instead",
+	})
+
+	rootCmd.AddCommand(&cobra.Command{
+		Use:   "list",
+		Short: "deprecated: use 'gh-not' instead",
+	})
+}
